pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-caching>=2.3.1",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "websocket-client>=1.8.0",
    "stravalib>=2.3",
    "requests>=2.32.3",
    "cryptography>=45.0.3",
    "streamlit>=1.45.1",
    "pandas>=2.3.0",
    "numpy>=2.2.6",
    "marshmallow>=4.0.0",
    "flask-jwt-extended>=4.7.1",
    "plotly>=6.1.2",
    "apscheduler>=3.11.0",
    "joblib>=1.5.1",
    "pytest>=8.4.0",
    "flask-socketio>=5.5.1",
    "flask-restx>=1.3.0",
    "scikit-learn>=1.7.0",
]
